+++ 1658525959877.sql
@@ -1,0 +1,16 @@
+-- Active: 1658524596278@@db.hijaeegxjbuivzckpijg.supabase.co@5432@postgres@public
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('1','Petirrojo','Merlot','vino tinto','27.000','1','1','245726');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('2','Posta Pizzela','Malbec','Vino Tinto','50.000','1','1','933404');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('3','Marques de villa de leyva','Cabernet Sauvignon','vino tinto','50.000','1','1','948930');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('4','Don melchor','Cabernet sauvignon','vino tinto','50.000','1','1','543152');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('5','Purple angel','Carmenere','vino tinto','50.000','1','1','330628');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('6','Muga','Tempranillo','vino tinto','50.000','2','1','692159');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('7','Chateu st michelle','Cabernet sauvignon','vino tinto','50.000','2','1','323324');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('8','Maison castel gran reserva','Cabernet sauvignon','vino tinto','50.000','2','1','191880');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('9','santa margherita','Pinot grigio','vino blanco','50.000','2','1','257609');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('10','Montes alpha M','80% Cabernet Sauvignon','vino tinto','50.000','2','1','782699');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('11','Angelica zapata','Malbec','vino tinto','50.000','3','1','850550');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('12','Gran enemigo','Malbec','vino tinto','50.000','3','1','652132');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('13','Zuccardi serie A','Malbec','vino tinto','50.000','3','1','438553');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('14','Cerro verde','Merlot','vino tinto','50.000','3','1','178631');
+insert into productos(`ID`,nombre,sepa,descripcion,precio,id_categoria,activo,producto) values('15','Codice','Tempranillo','vino tinto','50.000','3','1','265178');
