+++ config/database.php
@@ -1,24 +1,34 @@
 <?php
-class Database{
-    private $host = "localhost";
-    private $db_name = "tienda_online";
-    private $username = "tienda";
-    private $password = ")QZzqAiwP*3(EA4Q";
-    private $charset = "utf8";
-
-    function conectar(){
-        try{
-            $conexion = "mysql:host=".$this->host.";dbname=".$this->db_name.";charset=".$this->charset;
-            $options = [
-                PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION,
-                PDO::ATTR_EMULATE_PREPARES => false,
-            ];
-            $pdo = new PDO($conexion, $this->username, $this->password, $options);
-            return $pdo;
-        }catch(PDOException $e){
-            echo("Error connection: ".$e->getMessage());
-            exit;
-        }
+    // class Database{
+    //     public static function conectar(){
+    //         $host = "db.hijaeegxjbuivzckpijg.supabase.co";
+    //         $db_name = "postgres";
+    //         $port = "5432";
+    //         $username = "postgres";
+    //         $pasword = ")QZzqAiwP*3(EA4Q";
+    //         $charset = "utf8";
+    //         try{
+    //             $conect = new PDO("pgsql:host=$host; port=$port; dbname=$db_name", $username, $pasword);
+    //             echo "conectado";
+                
+    //         }catch(PDOException $exp){
+    //             echo "Error: " . $exp->getMessage();
+    //             exit;
+    //         }
+    //     }
+    // }
+    $host = "db.hijaeegxjbuivzckpijg.supabase.co";
+    $db_name = "postgres";
+    $port = "5432";
+    $username = "postgres";
+    $pasword = ")QZzqAiwP*3(EA4Q";
+    $charset = "utf8";
+    try{
+        $base_de_datos = new PDO("pgsql:host=$host; port=$port; dbname=$db_name", $username, $pasword);
+        echo "conectado";
+        $base_de_datos->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+    }catch(PDOException $exp){
+        echo "Error: " . $exp->getMessage();
+        exit;
     }
-}
 ?>
